test_twitscraper/requirements.txt
twitscraper
python-dotenv
